+++ src/django_components/components/__init__.py
@@ -1,4 +1,4 @@
-# NOTE: Components exported here are documented in
+# NOTE: Components exported here are documented in the API reference
 from django_components.components.dynamic import DynamicComponent
 
 __all__ = ["DynamicComponent"]